fix executor tests: remove impact from previous tests and to next tests: unload mindsdb module remove database path from environ
